Resources to estimate costs for a Cromwell workflow in GCP
Main relevant piece is scripts/estimate_billing.py which does the
transformation we want. To run that script, a directory of metadata
JSONs needs to be available, either locally or in a GCS bucket. This
directory is created via scripts/persist_artifacts.py which is made
available in manual-workflows case via the metadata-from-file tag in
start.sh and download_from_metadata call in server_startup.

helpers.sh updated to use persist_artifacts.py in its body for the
command of the same name.