Add resources to block access to Cromwell VMs outside of WashU VPN
One of the easiest security wins for us is locking the VM behind the
WashU VPN. We've done this before in the central-server approach, and
these same resources can be seen in the Terraform within that
directory. This commit simply adds them to the script shared between
GMS and manual-workflows approaches.

Tested manually that a simple workflow functions end-to-end with
`bam_to_fastq.wdl`. There were no issues with retrieving
packages (HTTP/HTTPS) which is the only thing I was worried about.

Concerns that aren't addressed here are:
- choosing to expose to Cromwell endpoints in some fashion
- additional security concerns beyond the VPN
- user-input names for network resources. Not sure we want this or not