Pull docker image version from a version controlled file
I was getting tired of trying to remember what the latest version
number was. Easy fix, store it in a file and just use that
file. git blame on the VERSION file _should_ imply repo at that state
is whats in the image. This only becomes true when we automate it
though, otherwise it's up to good practices.