Add option to convert cost JSON to CSV
Chris Miller wanted to see the task costs in a CSV format which is
much easier to parse than digging through nested JSON. The script
costs_json_to_csv.py accomplishes this. For convenience, added an
option to immediately do this in estimate_billing.py through the --csv
flag. Updated docs and Dockerfile to reflect changes