Remove ssh-access setting
The past few days, submitting workflows with this setting has resulted
in an error showing the SSH port 22 already bound. Removing this
setting resolves that, and worker nodes are still SSH accessible via
gcloud. I'm not sure how this happened exactly but all is back to
normal once removing this.